Changes for 1.1.0
The first version of trezenv include the bare minimum features to make it work as intended. Using click to manage cli commands. I added several alpha features such as set, allows to set the folder where trezenv will install. Check, to ensure paths are correct. And init as the entry point of the app.